error fixed

+++ migrations/000001_auth_service_table.up.sql
@@ -19,8 +19,8 @@ CREATE TABLE products
     category_id    UUID REFERENCES product_categories (id) NOT NULL,
     name           VARCHAR(50)                             NOT NULL,
     bill_format    VARCHAR(5)                              NOT NULL, -- можно заменить на ENUM, если есть ограниченное количество форматов
-    incoming_price DECIMAL(10, 2)                          NOT NULL,
-    standard_price DECIMAL(10, 2)                          NOT NULL,
+    incoming_price BIGINT                          NOT NULL,
+    standard_price BIGINT                          NOT NULL,
     total_count    INT       DEFAULT 0,
     created_by     UUID                                    NOT NULL,
     created_at     TIMESTAMP DEFAULT NOW()
@@ -32,7 +32,7 @@ CREATE TABLE sales
     id               UUID           DEFAULT gen_random_uuid() PRIMARY KEY,
     client_id        UUID           NOT NULL,
     sold_by          UUID           NOT NULL,
-    total_sale_price DECIMAL(10, 2) NOT NULL, -- общая сумма заказа
+    total_sale_price BIGINT NOT NULL, -- общая сумма заказа
     payment_method   payment_method DEFAULT 'uzs',
     created_at       TIMESTAMP      DEFAULT NOW()
 );
@@ -44,8 +44,8 @@ CREATE TABLE sales_items
     sale_id     UUID REFERENCES sales (id)    NOT NULL,
     product_id  UUID REFERENCES products (id) NOT NULL,
     quantity    INT  DEFAULT 1                NOT NULL,
-    sale_price  DECIMAL(10, 2)                NOT NULL,
-    total_price DECIMAL(10, 2)                NOT NULL -- общая цена за конкретный товар в заказе
+    sale_price  BIGINT                NOT NULL,
+    total_price BIGINT                NOT NULL -- общая цена за конкретный товар в заказе
 );
 
 -- Категории для учета денежных потоков
@@ -61,7 +61,7 @@ CREATE TABLE cash_flow
     id               UUID           DEFAULT gen_random_uuid() PRIMARY KEY,
     user_id          UUID                               NOT NULL,
     transaction_date TIMESTAMP      DEFAULT NOW(),
-    amount           DECIMAL(10, 2)                     NOT NULL,
+    amount           BIGINT                     NOT NULL,
     transaction_type transaction_type                   NOT NULL,
     category_id      UUID REFERENCES cash_category (id) NOT NULL,
     description      VARCHAR(255),
@@ -73,9 +73,9 @@ CREATE TABLE debts
 (
     id            UUID      DEFAULT gen_random_uuid() PRIMARY KEY,
     order_id      UUID REFERENCES sales (id) NOT NULL, -- Привязка долга к заказу
-    amount_paid   DECIMAL(10, 2)             NOT NULL, -- Сумма, уже оплаченная
-    amount_unpaid DECIMAL(10, 2)             NOT NULL, -- Сумма, остающаяся к оплате
-    total_debt    DECIMAL(10, 2)             NOT NULL, -- Общая сумма долга
+    amount_paid   BIGINT             NOT NULL, -- Сумма, уже оплаченная
+    amount_unpaid BIGINT             NOT NULL, -- Сумма, остающаяся к оплате
+    total_debt    BIGINT             NOT NULL, -- Общая сумма долга
     next_payment  DATE,
     last_paid_day TIMESTAMP DEFAULT NOW(),
     is_fully_paid BOOLEAN   DEFAULT FALSE,
@@ -89,7 +89,7 @@ CREATE TABLE debt_payments
     id           UUID      DEFAULT gen_random_uuid() PRIMARY KEY,
     debt_id      UUID REFERENCES debts (id) NOT NULL, -- Привязка к задолженности
     payment_date TIMESTAMP DEFAULT NOW(),
-    amount       DECIMAL(10, 2)             NOT NULL, -- Сумма частичного платежа
+    amount       BIGINT             NOT NULL, -- Сумма частичного платежа
     paid_by      UUID                       NOT NULL  -- Кто внес платёж
 );
 
@@ -99,7 +99,7 @@ CREATE TABLE purchases
     id             UUID           DEFAULT gen_random_uuid() PRIMARY KEY,
     supplier_id    UUID           NOT NULL,      -- Название поставщика или имя компании
     purchased_by   UUID           NOT NULL,      -- Кто произвел закупку
-    total_cost     DECIMAL(10, 2) NOT NULL,      -- Общая сумма закупки
+    total_cost     BIGINT NOT NULL,      -- Общая сумма закупки
     payment_method payment_method DEFAULT 'uzs', -- Способ оплаты
     description    TEXT,
     created_at     TIMESTAMP      DEFAULT NOW()  -- Время создания записи
@@ -112,6 +112,6 @@ CREATE TABLE purchase_items
     purchase_id    UUID REFERENCES purchases (id) NOT NULL, -- Ссылка на закупку
     product_id     UUID REFERENCES products (id)  NOT NULL, -- Ссылка на товар
     quantity       INT                            NOT NULL, -- Количество закупленного товара
-    purchase_price DECIMAL(10, 2)                 NOT NULL, -- Цена закупки за единицу товара
-    total_price    DECIMAL(10, 2)                 NOT NULL  -- Общая стоимость конкретного товара в закупке
+    purchase_price BIGINT                 NOT NULL, -- Цена закупки за единицу товара
+    total_price    BIGINT                 NOT NULL  -- Общая стоимость конкретного товара в закупке
 );
